refactor(Ladder): 사다리 게임 실행 로직 구현
