+++ src/main/java/ladder/domain/Result.java
@@ -2,54 +2,15 @@ package ladder.domain;
 
 import ladder.error.ErrorMessage;
 
-import java.util.*;
+import java.util.Collections;
+import java.util.LinkedHashMap;
+import java.util.Map;
 
 public class Result {
     private final Map<Name, Bet> result = new LinkedHashMap<>();
 
-    public Result(Names names, Bets bets) {
-        validate(names, bets);
-
-        for (int i = 0; i < names.size(); i++) {
-            result.put(names.getNames().get(i), bets.getBets().get(i));
-        }
-    }
-
-    private void validate(Names names, Bets bets) {
-        if (names.size() != bets.size())
-            throw new IllegalArgumentException(ErrorMessage.DIFFERENT_PARTICIPANTS_AND_BETS_COUNT.getMessage());
-    }
-
-    public void perform(Ladder ladder) {
-        List<Name> names = List.copyOf(result.keySet());
-        List<Bet> bets = new ArrayList<>(result.values());
-
-        for (int i = ladder.getLadder().size() - 1; i >= 0; i--) {
-            Line line = ladder.getLadder().get(i);
-            performByLine(line, bets);
-        }
-
-        update(names, bets);
-    }
-
-    private void performByLine(Line line, List<Bet> bets) {
-        for (int i = 0; i < line.size(); i++) {
-            swapBetIfExistLine(bets, i, line.existLineAtCell(i));
-        }
-    }
-
-    private void swapBetIfExistLine(List<Bet> bets, int index, boolean exist) {
-        if (!exist) {
-            return;
-        }
-
-        Collections.swap(bets, index, index + 1);
-    }
-
-    private void update(List<Name> names, List<Bet> bets) {
-        for (int i = 0; i < names.size(); i++) {
-            result.put(names.get(i), bets.get(i));
-        }
+    public Result(Map<Name, Bet> result) {
+        this.result.putAll(result);
     }
 
     public Bet getBetByName(Name name) {
